V1.0.2 beta.4

+++ Packages/BushelKit/Sources/BushelMachine/FileVersionSnapshotter/FileVersionSnapshotter.swift
@@ -1,6 +1,6 @@
 //
 // FileVersionSnapshotter.swift
-// Copyright (c) 2023 BrightDigit.
+// Copyright (c) 2024 BrightDigit.
 //
 
 #if os(macOS)
@@ -35,7 +35,9 @@
       let data = try JSON.encoder.encode(exportedConfiguration)
       let configurationFileURL = url.appendingPathComponent(URL.bushel.paths.machineJSONFileName)
       let newSnapshotsDirURL = url.appending(component: URL.bushel.paths.snapshotsDirectoryName)
-      try self.fileManager.removeItem(at: newSnapshotsDirURL)
+      if self.fileManager.directoryExists(at: newSnapshotsDirURL) == .directoryExists {
+        try self.fileManager.removeItem(at: newSnapshotsDirURL)
+      }
       try self.fileManager.createEmptyDirectory(
         at: newSnapshotsDirURL,
         withIntermediateDirectories: false,
